Fri, Nov 17, 2023, 9:41 PM +02:00

+++ filter-less/helpers.c
@@ -30,24 +30,30 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
 // Convert image to sepia
 
 
-// Helper function to cap the color values at 255
+/* Helper function to cap the color values at 255
 int cap(int value)
 {
     return value > 255 ? 255 : value;
 }
+*/
 
 // Convert image to sepia
 void sepia(int height, int width, RGBTRIPLE image[height][width])
 {
     // Iterate over each pixel in the image
+
+    //  comb through each row
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
         {
             // Calculate sepia values
-            int originalRed = image[i][j].rgbtRed;
-            int originalGreen = image[i][j].rgbtGreen;
-            int originalBlue = image[i][j].rgbtBlue;
+            // Convert pixels to float
+
+            float originalRed = image[i][j].rgbtRed;
+            float originalGreen = image[i][j].rgbtGreen;
+            float originalBlue = image[i][j].rgbtBlue;
+
 
             int sepiaRed = cap(0.393 * originalRed + 0.769 * originalGreen + 0.189 * originalBlue);
             int sepiaGreen = cap(0.349 * originalRed + 0.686 * originalGreen + 0.168 * originalBlue);
